+++ frontend/libmemio/libmemio.h
@@ -8,14 +8,14 @@ typedef struct memio {
 	bdbm_drv_info_t bdi;
 	bdbm_llm_req_t* rr;
 	int nr_punits;
-	size_t io_size; /* bytes */
-	size_t trim_size;
-	size_t trim_lbas;
+	uint64_t io_size; /* bytes */
+	uint64_t trim_size;
+	uint64_t trim_lbas;
 } memio_t;
 
 memio_t* memio_open ();
 void memio_wait (memio_t* mio);
-int memio_read (memio_t* mio, size_t lba, size_t len, uint8_t* data);
-int memio_write (memio_t* mio, size_t lba, size_t len, uint8_t* data);
-int memio_trim (memio_t* mio, size_t lba, size_t len);
+int memio_read (memio_t* mio, uint64_t lba, uint64_t len, uint8_t* data);
+int memio_write (memio_t* mio, uint64_t lba, uint64_t len, uint8_t* data);
+int memio_trim (memio_t* mio, uint64_t lba, uint64_t len);
 void memio_close (memio_t* mio);
